requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.1
python-dotenv==0.21.0
pydantic==1.10.4
httpx==0.24.0
python-multipart==0.0.6